.history/smartip/requirements_20221202110545.txt
django == 4.1.3
pillow
django_admin_interface
django_crispy_forms
crispy-bootstrap5
django-rest-framework
